Added solid lava bucket
Added the solid lava bucket to the game, added some recipes for the solid lava item, to make obsidian with a type of ice, and to make lava usig the bucket version if we heat it enough. solid lava bucket texture may or may not be logical as time of commiting.